backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pandas==2.1.4
chardet==5.2.0
# Platform transformation dependencies
redis==5.0.1
celery==5.3.4
polars==0.20.2
pyarrow==14.0.2
prometheus-client==0.19.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
# Task 20: Individual bonus statement generation
reportlab==4.0.8
xlsxwriter==3.1.9
